requirements.txt
fastapi
uvicorn[standard]
playwright
# google-generativeai # 더 이상 사용하지 않음
# langchain
# langchain-community, langchain-upstage 등 RAG에 필요한 구체적인 langchain 패키지 추가
langchain
langchain-community
langchain-upstage
faiss-cpu
langchain-cohere
langchain-experimental
langchain-groq  # Groq API 사용을 위해 추가
celery
redis
beautifulsoup4
python-multipart
python-dotenv # 환경 변수 관리를 위해 추가 